requirements.txt
pywhatkit
mysql-connector